DASHBOARD SAN SALVO/gfr-scada-web/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pandas
python-dotenv
httpx
PyJWT
requests
